requirements.txt
pillow
psycopg2-binary
py-cord[voice]
py-cord[speed]
python-dotenv
pytz
requests